Server/skillBadge/requirements.txt
Django==5.0.1
django_phonenumber_field==7.3.0
djangorestframework==3.14.0
